params

+++ sources/Configs.py
@@ -2,7 +2,7 @@ __author__ = 'giulio'
 
 
 class Configs:
-    floatType = 'float32'  # TODO move elsewhere
+    floatType = 'float64'  # TODO move elsewhere
     seed = 13
     output_dir = '/home/giulio/RNNs/models/'
